backend/requirements.txt
appdirs==1.4.0
argcomplete==1.8.2
astroid==2.0.4
beautifulsoup4==4.5.3
cached-property==1.3.0
chardet==2.3.0
Django==1.10.5
django-cors-headers==2.0.2
django-filter==1.0.1
django-rest-auth==0.9.0
django-url-filter==0.3.0
djangorestframework==3.5.3
djangorestframework-gis==0.13
docx2txt==0.6
EbookLib==0.16
enum34==1.1.6
isort==4.3.4
lazy-object-proxy==1.3.1
lxml==3.7.3
mccabe==0.6.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
psycopg2
pylint==2.1.1
pyparsing==2.1.10
python-pptx==0.6.5
six==1.10.0
SpeechRecognition==3.6.3
textract==1.5.0
typed-ast==1.1.0
Wand==0.4.4
wrapt==1.10.11
xlrd==1.0.0
XlsxWriter==0.9.6
ptvsd==4.3.2
